auth/requirements.txt
blinker==1.8.1
click==8.1.7
Flask==3.0.3
Flask-MySQLdb==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mysqlclient==2.2.4
PyJWT==2.8.0
Werkzeug==3.0.2
